+++ Examples/InvokableScripts/Package.swift
@@ -3,19 +3,19 @@
 import PackageDescription
 
 let package = Package(
-        name: "InvocableScripts",
+        name: "InvokableScripts",
         platforms: [
             .macOS(.v10_15)
         ],
         products: [
-            .executable(name: "invocable-scripts", targets: ["InvocableScripts"])
+            .executable(name: "invokable-scripts", targets: ["InvokableScripts"])
         ],
         dependencies: [
             .package(name: "influxdb-client-swift", path: "../../"),
             .package(url: "https://github.com/apple/swift-argument-parser", from: "1.1.2")
         ],
         targets: [
-            .executableTarget(name: "InvocableScripts", dependencies: [
+            .executableTarget(name: "InvokableScripts", dependencies: [
                 .product(name: "InfluxDBSwift", package: "influxdb-client-swift"),
                 .product(name: "ArgumentParser", package: "swift-argument-parser")
             ])